Adding swedish example
